reverted to very smooth animated style navigation on the front-end:...
In the mean time, I changed to do it in grails style, but later I realized, that it doesn't anymore scroll smoothly to the target,
but actually jumps to the subfolder /home/index.html#home. And also the animated line below the actual position would disappear.
This is not nice, so I reverted to my previous way. In main.gsp, it's still the 'grails-way', and it works all nicely.